+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 1.1.4
 - Created `array.iterate()` to iterate an `Array` using an `Iterator`
-- Created `array.iterator()` to return a `Generator` to iterate an `Array`, requires native Generators / from meta programming (no transpiling)
+- Created `array.iterator()` to return an `Iterator` for an `Array`
 
 ## 1.1.3
 - Moved more `variables` to `constants`